Polygon Drawings

+++ polylidar/glue.hpp
@@ -44,7 +44,7 @@ namespace polylidar {
         return extractPlanesAndPolygonsFromMesh(triangulation, config);
     }
 
-    std::vector<Polygon> _extractPolygonsFromMesh(py::array_t<double> vertices, py::array_t<double> triangles, py::array_t<double> halfedges,
+    std::vector<Polygon> _extractPolygonsFromMesh(py::array_t<double> vertices, py::array_t<size_t> triangles, py::array_t<size_t> halfedges,
                                                 double alpha = DEFAULT_ALPHA, double xyThresh = DEFAULT_XYTHRESH, double lmax=DEFAULT_LMAX, size_t minTriangles = DEFAULT_MINTRIANGLES,
                                                 size_t minHoleVertices = DEFAULT_MINHOLEVERTICES, double minBboxArea = DEFAULT_MINBBOX, double zThresh = DEFAULT_ZTHRESH,
                                                 double normThresh = DEFAULT_NORMTHRESH, double normThreshMin = DEFAULT_NORMTHRESH_MIN,
